+++ compiler/BIRReader.h
@@ -312,7 +312,7 @@ class ConstantPoolSet {
     BIRReader &readerRef = BIRReader::reader;
     ConstantPoolEntry *getEntry(int index) { return poolEntries[index].get(); }
     std::string getStringCp(uint32_t index);
-    uint32_t getIntCp(uint32_t index);
+    uint64_t getIntCp(uint32_t index);
     nballerina::Type getTypeCp(uint32_t index, bool voidToInt);
     float getFloatCp(uint32_t index);
     bool getBooleanCp(uint32_t index);
